refactor(android): lift whisperMetaText to shared ui/components
Hearth's local whisperMetaText helper had richer formatting than the
inline meta logic in WhispersScreen and InftDetail's Whispers panel:
- "*Tideborne* · whispered to · *Twilight* · 0042" (richer; Hearth)
- "*Tideborne* · 0042"                              (sparser; the rest)

Lifts the helper to ui/components/WhisperMeta.kt. WhispersScreen +
InftDetail Whispers panel adopt it. All three surfaces now show the
same comm meta formatting:

- whispers: "*from* · whispered to · *target* · tick"
- orch:     "*Orchestrator* · tick"
- human:    "*from* · steered · tick"

Pure refactor + de-duplication; no behavior change for Hearth.
