+++ apps/clan-world-mobile/app/src/main/java/world/clan/app/ui/screens/HearthScreen.kt
@@ -540,7 +540,7 @@ private fun WhispersList(
         else -> WhisperAccent.Default
       }
       WhisperRow(
-        meta = whisperMetaText(c),
+        meta = world.clan.app.ui.components.whisperMetaText(c),
         body = c.body,
         accent = accent,
       )
@@ -548,23 +548,5 @@ private fun WhispersList(
   }
 }
 
-@Composable
-private fun whisperMetaText(c: CombinedComm): AnnotatedString {
-  val from = c.fromClan?.let { clanDisplayName(it) }
-    ?: c.speaker
-    ?: when (c.kind) {
-      "orch" -> "Orchestrator"
-      "human" -> "Owner"
-      else -> "—"
-    }
-  val to = c.targetClan?.let { clanDisplayName(it) }
-  val tickStr = c.tick?.let { "%04d".format(it) } ?: "—"
-  val raw = when (c.kind) {
-    "whisper" -> if (to != null) "*$from* · whispered to · *$to* · $tickStr"
-                 else "*$from* · whispered · $tickStr"
-    "orch" -> "*Orchestrator* · $tickStr"
-    "human" -> "*$from* · steered · $tickStr"
-    else -> "*$from* · $tickStr"
-  }
-  return boldedMeta(raw.uppercase())
-}
+// whisperMetaText was lifted to ui.components.WhisperMeta so Hearth +
+// WhispersScreen + InftDetail all share the same comm-meta formatting.
