+++ apps/clan-world-mobile/app/src/main/java/world/clan/app/ui/screens/WhispersScreen.kt
@@ -276,14 +276,8 @@ private fun InboxRow(comm: CombinedComm) {
     "human" -> WhisperAccent.Ember
     else -> WhisperAccent.Default
   }
-  val from = comm.fromClan?.let { clanDisplayName(it) }
-    ?: comm.speaker
-    ?: when (comm.kind) {
-      "orch" -> "Orchestrator"; "human" -> "Owner"; else -> "—"
-    }
-  val tickStr = comm.tick?.let { "%04d".format(it) } ?: "—"
   WhisperRow(
-    meta = boldedMeta("*$from* · $tickStr".uppercase()),
+    meta = world.clan.app.ui.components.whisperMetaText(comm),
     body = comm.body,
     accent = accent,
   )
